@@ -1,100 +1,11 @@
 {
-  "_args": [
-    [
-      {
-        "raw": "bitcoinjs-lib",
-        "scope": null,
-        "escapedName": "bitcoinjs-lib",
-        "name": "bitcoinjs-lib",
-        "rawSpec": "",
-        "spec": "latest",
-        "type": "tag"
-      },
-      "/Users/coreyphillips/Documents/testing/bitcoinjslibTesting"
-    ]
-  ],
-  "_from": "bitcoinjs-lib@latest",
-  "_id": "bitcoinjs-lib@3.3.2",
-  "_inCache": true,
-  "_location": "/bitcoinjs-lib",
-  "_nodeVersion": "9.2.0",
-  "_npmOperationalInternal": {
-    "host": "s3://npm-registry-packages",
-    "tmp": "tmp/bitcoinjs-lib-3.3.2.tgz_1514030224360_0.2724244496785104"
-  },
-  "_npmUser": {
-    "name": "dcousens",
-    "email": "npm@dcousens.com"
-  },
-  "_npmVersion": "5.6.0",
-  "_phantomChildren": {},
-  "_requested": {
-    "raw": "bitcoinjs-lib",
-    "scope": null,
-    "escapedName": "bitcoinjs-lib",
-    "name": "bitcoinjs-lib",
-    "rawSpec": "",
-    "spec": "latest",
-    "type": "tag"
-  },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/bitcoinjs-lib/-/bitcoinjs-lib-3.3.2.tgz",
-  "_shasum": "780c9c53ecb1222adb463b58bef26386067b609a",
-  "_shrinkwrap": null,
-  "_spec": "bitcoinjs-lib",
-  "_where": "/Users/coreyphillips/Documents/testing/bitcoinjslibTesting",
-  "bugs": {
-    "url": "https://github.com/bitcoinjs/bitcoinjs-lib/issues"
-  },
-  "dependencies": {
-    "bech32": "^1.1.2",
-    "bigi": "^1.4.0",
-    "bip66": "^1.1.0",
-    "bitcoin-ops": "^1.3.0",
-    "bs58check": "^2.0.0",
-    "create-hash": "^1.1.0",
-    "create-hmac": "^1.1.3",
-    "ecurve": "^1.0.0",
-    "merkle-lib": "^2.0.10",
-    "pushdata-bitcoin": "^1.0.1",
-    "randombytes": "^2.0.1",
-    "safe-buffer": "^5.0.1",
-    "typeforce": "^1.11.3",
-    "varuint-bitcoin": "^1.0.4",
-    "wif": "^2.0.1"
-  },
+  "name": "bitcoinjs-lib",
+  "version": "4.0.1",
   "description": "Client-side Bitcoin JavaScript library",
-  "devDependencies": {
-    "async": "^2.0.1",
-    "bip39": "^2.3.0",
-    "bs58": "^4.0.0",
-    "cb-http-client": "^0.2.0",
-    "coinselect": "^3.1.1",
-    "dhttp": "^2.3.5",
-    "minimaldata": "^1.0.2",
-    "mocha": "^3.1.0",
-    "nyc": "^10.2.0",
-    "proxyquire": "^1.4.0",
-    "sinon": "^1.12.2",
-    "standard": "^9.0.2"
-  },
-  "directories": {},
-  "dist": {
-    "integrity": "sha512-l5qqvbaK8wwtANPf6oEffykycg4383XgEYdia1rI7/JpGf1jfRWlOUCvx5TiTZS7kyIvY4j/UhIQ2urLsvGkzw==",
-    "shasum": "780c9c53ecb1222adb463b58bef26386067b609a",
-    "tarball": "https://registry.npmjs.org/bitcoinjs-lib/-/bitcoinjs-lib-3.3.2.tgz"
-  },
+  "main": "./src/index.js",
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=8.0.0"
   },
-  "files": [
-    "src"
-  ],
-  "gitHead": "f4caaf42e7b58332d74c1540f88bcda7e55b82e6",
-  "homepage": "https://github.com/bitcoinjs/bitcoinjs-lib#readme",
   "keywords": [
     "bitcoinjs",
     "bitcoin",
