@@ -1,28 +1,28 @@
-var Buffer = require('safe-buffer').Buffer
-var bech32 = require('bech32')
-var bs58check = require('bs58check')
-var bscript = require('./script')
-var btemplates = require('./templates')
-var networks = require('./networks')
-var typeforce = require('typeforce')
-var types = require('./types')
+const Buffer = require('safe-buffer').Buffer
+const bech32 = require('bech32')
+const bs58check = require('bs58check')
+const bscript = require('./script')
+const networks = require('./networks')
+const typeforce = require('typeforce')
+const types = require('./types')
+const payments = require('./payments')
 
 function fromBase58Check (address) {
-  var payload = bs58check.decode(address)
+  const payload = bs58check.decode(address)
 
   // TODO: 4.0.0, move to "toOutputScript"
   if (payload.length < 21) throw new TypeError(address + ' is too short')
   if (payload.length > 21) throw new TypeError(address + ' is too long')
 
-  var version = payload.readUInt8(0)
-  var hash = payload.slice(1)
+  const version = payload.readUInt8(0)
+  const hash = payload.slice(1)
 
   return { version: version, hash: hash }
 }
 
 function fromBech32 (address) {
-  var result = bech32.decode(address)
-  var data = bech32.fromWords(result.words.slice(1))
+  const result = bech32.decode(address)
+  const data = bech32.fromWords(result.words.slice(1))
 
   return {
     version: result.words[0],
@@ -34,7 +34,7 @@ function fromBech32 (address) {
 function toBase58Check (hash, version) {
   typeforce(types.tuple(types.Hash160bit, types.UInt8), arguments)
 
-  var payload = Buffer.allocUnsafe(21)
+  const payload = Buffer.allocUnsafe(21)
   payload.writeUInt8(version, 0)
   hash.copy(payload, 1)
 
@@ -42,34 +42,34 @@ function toBase58Check (hash, version) {
 }
 
 function toBech32 (data, version, prefix) {
-  var words = bech32.toWords(data)
+  const words = bech32.toWords(data)
   words.unshift(version)
 
   return bech32.encode(prefix, words)
 }
 
-function fromOutputScript (outputScript, network) {
+function fromOutputScript (output, network) {
   network = network || networks.bitcoin
 
-  if (btemplates.pubKeyHash.output.check(outputScript)) return toBase58Check(bscript.compile(outputScript).slice(3, 23), network.pubKeyHash)
-  if (btemplates.scriptHash.output.check(outputScript)) return toBase58Check(bscript.compile(outputScript).slice(2, 22), network.scriptHash)
-  if (btemplates.witnessPubKeyHash.output.check(outputScript)) return toBech32(bscript.compile(outputScript).slice(2, 22), 0, network.bech32)
-  if (btemplates.witnessScriptHash.output.check(outputScript)) return toBech32(bscript.compile(outputScript).slice(2, 34), 0, network.bech32)
+  try { return payments.p2pkh({ output, network }).address } catch (e) {}
+  try { return payments.p2sh({ output, network }).address } catch (e) {}
+  try { return payments.p2wpkh({ output, network }).address } catch (e) {}
+  try { return payments.p2wsh({ output, network }).address } catch (e) {}
 
-  throw new Error(bscript.toASM(outputScript) + ' has no matching Address')
+  throw new Error(bscript.toASM(output) + ' has no matching Address')
 }
 
 function toOutputScript (address, network) {
   network = network || networks.bitcoin
 
-  var decode
+  let decode
   try {
     decode = fromBase58Check(address)
   } catch (e) {}
 
   if (decode) {
-    if (decode.version === network.pubKeyHash) return btemplates.pubKeyHash.output.encode(decode.hash)
-    if (decode.version === network.scriptHash) return btemplates.scriptHash.output.encode(decode.hash)
+    if (decode.version === network.pubKeyHash) return payments.p2pkh({ hash: decode.hash }).output
+    if (decode.version === network.scriptHash) return payments.p2sh({ hash: decode.hash }).output
   } else {
     try {
       decode = fromBech32(address)
@@ -78,8 +78,8 @@ function toOutputScript (address, network) {
     if (decode) {
       if (decode.prefix !== network.bech32) throw new Error(address + ' has an invalid prefix')
       if (decode.version === 0) {
-        if (decode.data.length === 20) return btemplates.witnessPubKeyHash.output.encode(decode.data)
-        if (decode.data.length === 32) return btemplates.witnessScriptHash.output.encode(decode.data)
+        if (decode.data.length === 20) return payments.p2wpkh({ hash: decode.data }).output
+        if (decode.data.length === 32) return payments.p2wsh({ hash: decode.data }).output
       }
     }
   }
