@@ -1,12 +1,11 @@
-var Buffer = require('safe-buffer').Buffer
-var bcrypto = require('./crypto')
-var fastMerkleRoot = require('merkle-lib/fastRoot')
-var typeforce = require('typeforce')
-var types = require('./types')
-var varuint = require('varuint-bitcoin')
-var bufferReverse = require('buffer-reverse')
+const Buffer = require('safe-buffer').Buffer
+const bcrypto = require('./crypto')
+const fastMerkleRoot = require('merkle-lib/fastRoot')
+const typeforce = require('typeforce')
+const types = require('./types')
+const varuint = require('varuint-bitcoin')
 
-var Transaction = require('./transaction')
+const Transaction = require('./transaction')
 
 function Block () {
   this.version = 1
@@ -20,25 +19,25 @@ function Block () {
 Block.fromBuffer = function (buffer) {
   if (buffer.length < 80) throw new Error('Buffer too small (< 80 bytes)')
 
-  var offset = 0
+  let offset = 0
   function readSlice (n) {
     offset += n
     return buffer.slice(offset - n, offset)
   }
 
   function readUInt32 () {
-    var i = buffer.readUInt32LE(offset)
+    const i = buffer.readUInt32LE(offset)
     offset += 4
     return i
   }
 
   function readInt32 () {
-    var i = buffer.readInt32LE(offset)
+    const i = buffer.readInt32LE(offset)
     offset += 4
     return i
   }
 
-  var block = new Block()
+  const block = new Block()
   block.version = readInt32()
   block.prevHash = readSlice(32)
   block.merkleRoot = readSlice(32)
@@ -49,22 +48,22 @@ Block.fromBuffer = function (buffer) {
   if (buffer.length === 80) return block
 
   function readVarInt () {
-    var vi = varuint.decode(buffer, offset)
+    const vi = varuint.decode(buffer, offset)
     offset += varuint.decode.bytes
     return vi
   }
 
   function readTransaction () {
-    var tx = Transaction.fromBuffer(buffer.slice(offset), true)
+    const tx = Transaction.fromBuffer(buffer.slice(offset), true)
     offset += tx.byteLength()
     return tx
   }
 
-  var nTransactions = readVarInt()
+  const nTransactions = readVarInt()
   block.transactions = []
 
   for (var i = 0; i < nTransactions; ++i) {
-    var tx = readTransaction()
+    const tx = readTransaction()
     block.transactions.push(tx)
   }
 
@@ -88,11 +87,11 @@ Block.prototype.getHash = function () {
 }
 
 Block.prototype.getId = function () {
-  return bufferReverse(new Buffer(this.getHash(), 'hex'))
+  return this.getHash().reverse().toString('hex')
 }
 
 Block.prototype.getUTCDate = function () {
-  var date = new Date(0) // epoch
+  const date = new Date(0) // epoch
   date.setUTCSeconds(this.timestamp)
 
   return date
@@ -100,9 +99,9 @@ Block.prototype.getUTCDate = function () {
 
 // TODO: buffer, offset compatibility
 Block.prototype.toBuffer = function (headersOnly) {
-  var buffer = Buffer.allocUnsafe(this.byteLength(headersOnly))
+  const buffer = Buffer.allocUnsafe(this.byteLength(headersOnly))
 
-  var offset = 0
+  let offset = 0
   function writeSlice (slice) {
     slice.copy(buffer, offset)
     offset += slice.length
@@ -130,7 +129,7 @@ Block.prototype.toBuffer = function (headersOnly) {
   offset += varuint.encode.bytes
 
   this.transactions.forEach(function (tx) {
-    var txSize = tx.byteLength() // TODO: extract from toBuffer?
+    const txSize = tx.byteLength() // TODO: extract from toBuffer?
     tx.toBuffer(buffer, offset)
     offset += txSize
   })
@@ -143,9 +142,9 @@ Block.prototype.toHex = function (headersOnly) {
 }
 
 Block.calculateTarget = function (bits) {
-  var exponent = ((bits & 0xff000000) >> 24) - 3
-  var mantissa = bits & 0x007fffff
-  var target = Buffer.alloc(32, 0)
+  const exponent = ((bits & 0xff000000) >> 24) - 3
+  const mantissa = bits & 0x007fffff
+  const target = Buffer.alloc(32, 0)
   target.writeUInt32BE(mantissa, 28 - exponent)
   return target
 }
@@ -154,7 +153,7 @@ Block.calculateMerkleRoot = function (transactions) {
   typeforce([{ getHash: types.Function }], transactions)
   if (transactions.length === 0) throw TypeError('Cannot compute merkle root for zero transactions')
 
-  var hashes = transactions.map(function (transaction) {
+  const hashes = transactions.map(function (transaction) {
     return transaction.getHash()
   })
 
@@ -164,13 +163,13 @@ Block.calculateMerkleRoot = function (transactions) {
 Block.prototype.checkMerkleRoot = function () {
   if (!this.transactions) return false
 
-  var actualMerkleRoot = Block.calculateMerkleRoot(this.transactions)
+  const actualMerkleRoot = Block.calculateMerkleRoot(this.transactions)
   return this.merkleRoot.compare(actualMerkleRoot) === 0
 }
 
 Block.prototype.checkProofOfWork = function () {
-  var hash = bufferReverse(new Buffer(this.getHash()))
-  var target = Block.calculateTarget(this.bits)
+  const hash = this.getHash().reverse()
+  const target = Block.calculateTarget(this.bits)
 
   return hash.compare(target) <= 0
 }
