@@ -1,23 +1,16 @@
-var script = require('./script')
-
-var templates = require('./templates')
-for (var key in templates) {
-  script[key] = templates[key]
-}
+const script = require('./script')
 
 module.exports = {
-  bufferutils: require('./bufferutils'), // TODO: remove in 4.0.0
-
   Block: require('./block'),
   ECPair: require('./ecpair'),
-  ECSignature: require('./ecsignature'),
-  HDNode: require('./hdnode'),
   Transaction: require('./transaction'),
   TransactionBuilder: require('./transaction_builder'),
 
   address: require('./address'),
+  bip32: require('bip32'),
   crypto: require('./crypto'),
   networks: require('./networks'),
   opcodes: require('bitcoin-ops'),
+  payments: require('./payments'),
   script: script
 }
