@@ -1,10 +1,10 @@
-var Buffer = require('safe-buffer').Buffer
+const Buffer = require('safe-buffer').Buffer
 
 function decode (buffer, maxLength, minimal) {
   maxLength = maxLength || 4
   minimal = minimal === undefined ? true : minimal
 
-  var length = buffer.length
+  const length = buffer.length
   if (length === 0) return 0
   if (length > maxLength) throw new TypeError('Script number overflow')
   if (minimal) {
@@ -15,16 +15,15 @@ function decode (buffer, maxLength, minimal) {
 
   // 40-bit
   if (length === 5) {
-    var a = buffer.readUInt32LE(0)
-    var b = buffer.readUInt8(4)
+    const a = buffer.readUInt32LE(0)
+    const b = buffer.readUInt8(4)
 
     if (b & 0x80) return -(((b & ~0x80) * 0x100000000) + a)
     return (b * 0x100000000) + a
   }
 
-  var result = 0
-
   // 32-bit / 24-bit / 16-bit / 8-bit
+  let result = 0
   for (var i = 0; i < length; ++i) {
     result |= buffer[i] << (8 * i)
   }
@@ -35,18 +34,18 @@ function decode (buffer, maxLength, minimal) {
 
 function scriptNumSize (i) {
   return i > 0x7fffffff ? 5
-  : i > 0x7fffff ? 4
-  : i > 0x7fff ? 3
-  : i > 0x7f ? 2
-  : i > 0x00 ? 1
-  : 0
+    : i > 0x7fffff ? 4
+      : i > 0x7fff ? 3
+        : i > 0x7f ? 2
+          : i > 0x00 ? 1
+            : 0
 }
 
 function encode (number) {
-  var value = Math.abs(number)
-  var size = scriptNumSize(value)
-  var buffer = Buffer.allocUnsafe(size)
-  var negative = number < 0
+  let value = Math.abs(number)
+  const size = scriptNumSize(value)
+  const buffer = Buffer.allocUnsafe(size)
+  const negative = number < 0
 
   for (var i = 0; i < size; ++i) {
     buffer.writeUInt8(value & 0xff, i)
