@@ -1,12 +1,10 @@
 // OP_HASH160 {scriptHash} OP_EQUAL
 
-var bscript = require('../../script')
-var types = require('../../types')
-var typeforce = require('typeforce')
-var OPS = require('bitcoin-ops')
+const bscript = require('../../script')
+const OPS = require('bitcoin-ops')
 
 function check (script) {
-  var buffer = bscript.compile(script)
+  const buffer = bscript.compile(script)
 
   return buffer.length === 23 &&
     buffer[0] === OPS.OP_HASH160 &&
@@ -15,20 +13,4 @@ function check (script) {
 }
 check.toJSON = function () { return 'scriptHash output' }
 
-function encode (scriptHash) {
-  typeforce(types.Hash160bit, scriptHash)
-
-  return bscript.compile([OPS.OP_HASH160, scriptHash, OPS.OP_EQUAL])
-}
-
-function decode (buffer) {
-  typeforce(check, buffer)
-
-  return buffer.slice(2, 22)
-}
-
-module.exports = {
-  check: check,
-  decode: decode,
-  encode: encode
-}
+module.exports = { check }
