@@ -1,15 +1,15 @@
 // OP_RETURN {aa21a9ed} {commitment}
 
-var Buffer = require('safe-buffer').Buffer
-var bscript = require('../../script')
-var types = require('../../types')
-var typeforce = require('typeforce')
-var OPS = require('bitcoin-ops')
+const Buffer = require('safe-buffer').Buffer
+const bscript = require('../../script')
+const types = require('../../types')
+const typeforce = require('typeforce')
+const OPS = require('bitcoin-ops')
 
-var HEADER = Buffer.from('aa21a9ed', 'hex')
+const HEADER = Buffer.from('aa21a9ed', 'hex')
 
 function check (script) {
-  var buffer = bscript.compile(script)
+  const buffer = bscript.compile(script)
 
   return buffer.length > 37 &&
     buffer[0] === OPS.OP_RETURN &&
@@ -22,7 +22,7 @@ check.toJSON = function () { return 'Witness commitment output' }
 function encode (commitment) {
   typeforce(types.Hash256bit, commitment)
 
-  var buffer = Buffer.allocUnsafe(36)
+  const buffer = Buffer.allocUnsafe(36)
   HEADER.copy(buffer, 0)
   commitment.copy(buffer, 4)
 
