@@ -1,12 +1,10 @@
 // OP_0 {pubKeyHash}
 
-var bscript = require('../../script')
-var types = require('../../types')
-var typeforce = require('typeforce')
-var OPS = require('bitcoin-ops')
+const bscript = require('../../script')
+const OPS = require('bitcoin-ops')
 
 function check (script) {
-  var buffer = bscript.compile(script)
+  const buffer = bscript.compile(script)
 
   return buffer.length === 22 &&
     buffer[0] === OPS.OP_0 &&
@@ -14,20 +12,6 @@ function check (script) {
 }
 check.toJSON = function () { return 'Witness pubKeyHash output' }
 
-function encode (pubKeyHash) {
-  typeforce(types.Hash160bit, pubKeyHash)
-
-  return bscript.compile([OPS.OP_0, pubKeyHash])
-}
-
-function decode (buffer) {
-  typeforce(check, buffer)
-
-  return buffer.slice(2)
-}
-
 module.exports = {
-  check: check,
-  decode: decode,
-  encode: encode
+  check
 }
