@@ -1,12 +1,10 @@
 // OP_0 {scriptHash}
 
-var bscript = require('../../script')
-var types = require('../../types')
-var typeforce = require('typeforce')
-var OPS = require('bitcoin-ops')
+const bscript = require('../../script')
+const OPS = require('bitcoin-ops')
 
 function check (script) {
-  var buffer = bscript.compile(script)
+  const buffer = bscript.compile(script)
 
   return buffer.length === 34 &&
     buffer[0] === OPS.OP_0 &&
@@ -14,20 +12,4 @@ function check (script) {
 }
 check.toJSON = function () { return 'Witness scriptHash output' }
 
-function encode (scriptHash) {
-  typeforce(types.Hash256bit, scriptHash)
-
-  return bscript.compile([OPS.OP_0, scriptHash])
-}
-
-function decode (buffer) {
-  typeforce(check, buffer)
-
-  return buffer.slice(2)
-}
-
-module.exports = {
-  check: check,
-  decode: decode,
-  encode: encode
-}
+module.exports = { check }
