@@ -1,189 +1,140 @@
-var Buffer = require('safe-buffer').Buffer
-var baddress = require('./address')
-var bcrypto = require('./crypto')
-var bscript = require('./script')
-var bufferReverse = require('buffer-reverse')
-var btemplates = require('./templates')
-var networks = require('./networks')
-var ops = require('bitcoin-ops')
-var typeforce = require('typeforce')
-var types = require('./types')
-var scriptTypes = btemplates.types
-var SIGNABLE = [btemplates.types.P2PKH, btemplates.types.P2PK, btemplates.types.MULTISIG]
-var P2SH = SIGNABLE.concat([btemplates.types.P2WPKH, btemplates.types.P2WSH])
+const Buffer = require('safe-buffer').Buffer
+const baddress = require('./address')
+const bcrypto = require('./crypto')
+const bscript = require('./script')
+const networks = require('./networks')
+const ops = require('bitcoin-ops')
+const payments = require('./payments')
+const typeforce = require('typeforce')
+const types = require('./types')
+const classify = require('./classify')
+const bufferReverse = require('buffer-reverse')
+const SCRIPT_TYPES = classify.types
 
-var ECPair = require('./ecpair')
-var ECSignature = require('./ecsignature')
-var Transaction = require('./transaction')
+const ECPair = require('./ecpair')
+const Transaction = require('./transaction')
 
-function supportedType (type) {
-  return SIGNABLE.indexOf(type) !== -1
-}
+function expandInput (scriptSig, witnessStack, type, scriptPubKey) {
+  if (scriptSig.length === 0 && witnessStack.length === 0) return {}
+  if (!type) {
+    let ssType = classify.input(scriptSig, true)
+    let wsType = classify.witness(witnessStack, true)
+    if (ssType === SCRIPT_TYPES.NONSTANDARD) ssType = undefined
+    if (wsType === SCRIPT_TYPES.NONSTANDARD) wsType = undefined
+    type = ssType || wsType
+  }
 
-function supportedP2SHType (type) {
-  return P2SH.indexOf(type) !== -1
-}
-
-function extractChunks (type, chunks, script) {
-  var pubKeys = []
-  var signatures = []
   switch (type) {
-    case scriptTypes.P2PKH:
-      // if (redeemScript) throw new Error('Nonstandard... P2SH(P2PKH)')
-      pubKeys = chunks.slice(1)
-      signatures = chunks.slice(0, 1)
-      break
+    case SCRIPT_TYPES.P2WPKH: {
+      const { output, pubkey, signature } = payments.p2wpkh({ witness: witnessStack })
 
-    case scriptTypes.P2PK:
-      pubKeys[0] = script ? btemplates.pubKey.output.decode(script) : undefined
-      signatures = chunks.slice(0, 1)
-      break
-
-    case scriptTypes.MULTISIG:
-      if (script) {
-        var multisig = btemplates.multisig.output.decode(script)
-        pubKeys = multisig.pubKeys
+      return {
+        prevOutScript: output,
+        prevOutType: SCRIPT_TYPES.P2WPKH,
+        pubkeys: [pubkey],
+        signatures: [signature]
       }
+    }
 
-      signatures = chunks.slice(1).map(function (chunk) {
-        return chunk.length === 0 ? undefined : chunk
-      })
-      break
+    case SCRIPT_TYPES.P2PKH: {
+      const { output, pubkey, signature } = payments.p2pkh({ input: scriptSig })
+
+      return {
+        prevOutScript: output,
+        prevOutType: SCRIPT_TYPES.P2PKH,
+        pubkeys: [pubkey],
+        signatures: [signature]
+      }
+    }
+
+    case SCRIPT_TYPES.P2PK: {
+      const { signature } = payments.p2pk({ input: scriptSig })
+
+      return {
+        prevOutType: SCRIPT_TYPES.P2PK,
+        pubkeys: [undefined],
+        signatures: [signature]
+      }
+    }
+
+    case SCRIPT_TYPES.MULTISIG: {
+      const { pubkeys, signatures } = payments.p2ms({
+        input: scriptSig,
+        output: scriptPubKey
+      }, { allowIncomplete: true })
+
+      return {
+        prevOutType: SCRIPT_TYPES.MULTISIG,
+        pubkeys: pubkeys,
+        signatures: signatures
+      }
+    }
+  }
+
+  if (type === SCRIPT_TYPES.P2SH) {
+    const { output, redeem } = payments.p2sh({
+      input: scriptSig,
+      witness: witnessStack
+    })
+
+    const outputType = classify.output(redeem.output)
+    const expanded = expandInput(redeem.input, redeem.witness, outputType, redeem.output)
+    if (!expanded.prevOutType) return {}
+
+    return {
+      prevOutScript: output,
+      prevOutType: SCRIPT_TYPES.P2SH,
+      redeemScript: redeem.output,
+      redeemScriptType: expanded.prevOutType,
+      witnessScript: expanded.witnessScript,
+      witnessScriptType: expanded.witnessScriptType,
+
+      pubkeys: expanded.pubkeys,
+      signatures: expanded.signatures
+    }
+  }
+
+  if (type === SCRIPT_TYPES.P2WSH) {
+    const { output, redeem } = payments.p2wsh({
+      input: scriptSig,
+      witness: witnessStack
+    })
+    const outputType = classify.output(redeem.output)
+    let expanded
+    if (outputType === SCRIPT_TYPES.P2WPKH) {
+      expanded = expandInput(redeem.input, redeem.witness, outputType)
+    } else {
+      expanded = expandInput(bscript.compile(redeem.witness), [], outputType, redeem.output)
+    }
+    if (!expanded.prevOutType) return {}
+
+    return {
+      prevOutScript: output,
+      prevOutType: SCRIPT_TYPES.P2WSH,
+      witnessScript: redeem.output,
+      witnessScriptType: expanded.prevOutType,
+
+      pubkeys: expanded.pubkeys,
+      signatures: expanded.signatures
+    }
   }
 
   return {
-    pubKeys: pubKeys,
-    signatures: signatures
+    prevOutType: SCRIPT_TYPES.NONSTANDARD,
+    prevOutScript: scriptSig
   }
-}
-function expandInput (scriptSig, witnessStack) {
-  if (scriptSig.length === 0 && witnessStack.length === 0) return {}
-
-  var prevOutScript
-  var prevOutType
-  var scriptType
-  var script
-  var redeemScript
-  var witnessScript
-  var witnessScriptType
-  var redeemScriptType
-  var witness = false
-  var p2wsh = false
-  var p2sh = false
-  var witnessProgram
-  var chunks
-
-  var scriptSigChunks = bscript.decompile(scriptSig)
-  var sigType = btemplates.classifyInput(scriptSigChunks, true)
-  if (sigType === scriptTypes.P2SH) {
-    p2sh = true
-    redeemScript = scriptSigChunks[scriptSigChunks.length - 1]
-    redeemScriptType = btemplates.classifyOutput(redeemScript)
-    prevOutScript = btemplates.scriptHash.output.encode(bcrypto.hash160(redeemScript))
-    prevOutType = scriptTypes.P2SH
-    script = redeemScript
-  }
-
-  var classifyWitness = btemplates.classifyWitness(witnessStack, true)
-  if (classifyWitness === scriptTypes.P2WSH) {
-    witnessScript = witnessStack[witnessStack.length - 1]
-    witnessScriptType = btemplates.classifyOutput(witnessScript)
-    p2wsh = true
-    witness = true
-    if (scriptSig.length === 0) {
-      prevOutScript = btemplates.witnessScriptHash.output.encode(bcrypto.sha256(witnessScript))
-      prevOutType = scriptTypes.P2WSH
-      if (redeemScript !== undefined) {
-        throw new Error('Redeem script given when unnecessary')
-      }
-      // bare witness
-    } else {
-      if (!redeemScript) {
-        throw new Error('No redeemScript provided for P2WSH, but scriptSig non-empty')
-      }
-      witnessProgram = btemplates.witnessScriptHash.output.encode(bcrypto.sha256(witnessScript))
-      if (!redeemScript.equals(witnessProgram)) {
-        throw new Error('Redeem script didn\'t match witnessScript')
-      }
-    }
-
-    if (!supportedType(btemplates.classifyOutput(witnessScript))) {
-      throw new Error('unsupported witness script')
-    }
-
-    script = witnessScript
-    scriptType = witnessScriptType
-    chunks = witnessStack.slice(0, -1)
-  } else if (classifyWitness === scriptTypes.P2WPKH) {
-    witness = true
-    var key = witnessStack[witnessStack.length - 1]
-    var keyHash = bcrypto.hash160(key)
-    if (scriptSig.length === 0) {
-      prevOutScript = btemplates.witnessPubKeyHash.output.encode(keyHash)
-      prevOutType = scriptTypes.P2WPKH
-      if (typeof redeemScript !== 'undefined') {
-        throw new Error('Redeem script given when unnecessary')
-      }
-    } else {
-      if (!redeemScript) {
-        throw new Error('No redeemScript provided for P2WPKH, but scriptSig wasn\'t empty')
-      }
-      witnessProgram = btemplates.witnessPubKeyHash.output.encode(keyHash)
-      if (!redeemScript.equals(witnessProgram)) {
-        throw new Error('Redeem script did not have the right witness program')
-      }
-    }
-
-    scriptType = scriptTypes.P2PKH
-    chunks = witnessStack
-  } else if (redeemScript) {
-    if (!supportedP2SHType(redeemScriptType)) {
-      throw new Error('Bad redeemscript!')
-    }
-
-    script = redeemScript
-    scriptType = redeemScriptType
-    chunks = scriptSigChunks.slice(0, -1)
-  } else {
-    prevOutType = scriptType = btemplates.classifyInput(scriptSig)
-    chunks = scriptSigChunks
-  }
-
-  var expanded = extractChunks(scriptType, chunks, script)
-
-  var result = {
-    pubKeys: expanded.pubKeys,
-    signatures: expanded.signatures,
-    prevOutScript: prevOutScript,
-    prevOutType: prevOutType,
-    signType: scriptType,
-    signScript: script,
-    witness: Boolean(witness)
-  }
-
-  if (p2sh) {
-    result.redeemScript = redeemScript
-    result.redeemScriptType = redeemScriptType
-  }
-
-  if (p2wsh) {
-    result.witnessScript = witnessScript
-    result.witnessScriptType = witnessScriptType
-  }
-
-  return result
 }
 
 // could be done in expandInput, but requires the original Transaction for hashForSignature
 function fixMultisigOrder (input, transaction, vin) {
-  if (input.redeemScriptType !== scriptTypes.MULTISIG || !input.redeemScript) return
-  if (input.pubKeys.length === input.signatures.length) return
+  if (input.redeemScriptType !== SCRIPT_TYPES.MULTISIG || !input.redeemScript) return
+  if (input.pubkeys.length === input.signatures.length) return
 
-  var unmatched = input.signatures.concat()
+  const unmatched = input.signatures.concat()
 
-  input.signatures = input.pubKeys.map(function (pubKey) {
-    var keyPair = ECPair.fromPublicKeyBuffer(pubKey)
-    var match
+  input.signatures = input.pubkeys.map(function (pubKey) {
+    const keyPair = ECPair.fromPublicKey(pubKey)
+    let match
 
     // check for a signature
     unmatched.some(function (signature, i) {
@@ -191,8 +142,8 @@ function fixMultisigOrder (input, transaction, vin) {
       if (!signature) return false
 
       // TODO: avoid O(n) hashForSignature
-      var parsed = ECSignature.parseScriptSignature(signature)
-      var hash = transaction.hashForSignature(vin, input.redeemScript, parsed.hashType)
+      const parsed = bscript.signature.decode(signature)
+      const hash = transaction.hashForSignature(vin, input.redeemScript, parsed.hashType)
 
       // skip if signature does not match pubKey
       if (!keyPair.verify(hash, parsed.signature)) return false
@@ -208,279 +159,293 @@ function fixMultisigOrder (input, transaction, vin) {
   })
 }
 
-function expandOutput (script, scriptType, ourPubKey) {
+function expandOutput (script, ourPubKey) {
   typeforce(types.Buffer, script)
+  const type = classify.output(script)
 
-  var scriptChunks = bscript.decompile(script)
-  if (!scriptType) {
-    scriptType = btemplates.classifyOutput(script)
+  switch (type) {
+    case SCRIPT_TYPES.P2PKH: {
+      if (!ourPubKey) return { type }
+
+      // does our hash160(pubKey) match the output scripts?
+      const pkh1 = payments.p2pkh({ output: script }).hash
+      const pkh2 = bcrypto.hash160(ourPubKey)
+      if (!pkh1.equals(pkh2)) return { type }
+
+      return {
+        type,
+        pubkeys: [ourPubKey],
+        signatures: [undefined]
+      }
+    }
+
+    case SCRIPT_TYPES.P2WPKH: {
+      if (!ourPubKey) return { type }
+
+      // does our hash160(pubKey) match the output scripts?
+      const wpkh1 = payments.p2wpkh({ output: script }).hash
+      const wpkh2 = bcrypto.hash160(ourPubKey)
+      if (!wpkh1.equals(wpkh2)) return { type }
+
+      return {
+        type,
+        pubkeys: [ourPubKey],
+        signatures: [undefined]
+      }
+    }
+
+    case SCRIPT_TYPES.P2PK: {
+      const p2pk = payments.p2pk({ output: script })
+      return {
+        type,
+        pubkeys: [p2pk.pubkey],
+        signatures: [undefined]
+      }
+    }
+
+    case SCRIPT_TYPES.MULTISIG: {
+      const p2ms = payments.p2ms({ output: script })
+      return {
+        type,
+        pubkeys: p2ms.pubkeys,
+        signatures: p2ms.pubkeys.map(() => undefined)
+      }
+    }
   }
 
-  var pubKeys = []
-
-  switch (scriptType) {
-    // does our hash160(pubKey) match the output scripts?
-    case scriptTypes.P2PKH:
-      if (!ourPubKey) break
-
-      var pkh1 = scriptChunks[2]
-      var pkh2 = bcrypto.hash160(ourPubKey)
-      if (pkh1.equals(pkh2)) pubKeys = [ourPubKey]
-      break
-
-    // does our hash160(pubKey) match the output scripts?
-    case scriptTypes.P2WPKH:
-      if (!ourPubKey) break
-
-      var wpkh1 = scriptChunks[1]
-      var wpkh2 = bcrypto.hash160(ourPubKey)
-      if (wpkh1.equals(wpkh2)) pubKeys = [ourPubKey]
-      break
-
-    case scriptTypes.P2PK:
-      pubKeys = scriptChunks.slice(0, 1)
-      break
-
-    case scriptTypes.MULTISIG:
-      pubKeys = scriptChunks.slice(1, -2)
-      break
-
-    default: return { scriptType: scriptType }
-  }
-
-  return {
-    pubKeys: pubKeys,
-    scriptType: scriptType,
-    signatures: pubKeys.map(function () { return undefined })
-  }
+  return { type }
 }
 
-function checkP2SHInput (input, redeemScriptHash) {
-  if (input.prevOutType) {
-    if (input.prevOutType !== scriptTypes.P2SH) throw new Error('PrevOutScript must be P2SH')
-
-    var prevOutScriptScriptHash = bscript.decompile(input.prevOutScript)[1]
-    if (!prevOutScriptScriptHash.equals(redeemScriptHash)) throw new Error('Inconsistent hash160(RedeemScript)')
-  }
-}
-
-function checkP2WSHInput (input, witnessScriptHash) {
-  if (input.prevOutType) {
-    if (input.prevOutType !== scriptTypes.P2WSH) throw new Error('PrevOutScript must be P2WSH')
-
-    var scriptHash = bscript.decompile(input.prevOutScript)[1]
-    if (!scriptHash.equals(witnessScriptHash)) throw new Error('Inconsistent sha25(WitnessScript)')
-  }
-}
-
-function prepareInput (input, kpPubKey, redeemScript, witnessValue, witnessScript) {
-  var expanded
-  var prevOutType
-  var prevOutScript
-
-  var p2sh = false
-  var p2shType
-  var redeemScriptHash
-
-  var witness = false
-  var p2wsh = false
-  var witnessType
-  var witnessScriptHash
-
-  var signType
-  var signScript
-
+function prepareInput (input, ourPubKey, redeemScript, witnessValue, witnessScript) {
   if (redeemScript && witnessScript) {
-    redeemScriptHash = bcrypto.hash160(redeemScript)
-    witnessScriptHash = bcrypto.sha256(witnessScript)
-    checkP2SHInput(input, redeemScriptHash)
+    const p2wsh = payments.p2wsh({ redeem: { output: witnessScript } })
+    const p2wshAlt = payments.p2wsh({ output: redeemScript })
+    const p2sh = payments.p2sh({ redeem: { output: redeemScript } })
+    const p2shAlt = payments.p2sh({ redeem: p2wsh })
 
-    if (!redeemScript.equals(btemplates.witnessScriptHash.output.encode(witnessScriptHash))) throw new Error('Witness script inconsistent with redeem script')
+    // enforces P2SH(P2WSH(...))
+    if (!p2wsh.hash.equals(p2wshAlt.hash)) throw new Error('Witness script inconsistent with prevOutScript')
+    if (!p2sh.hash.equals(p2shAlt.hash)) throw new Error('Redeem script inconsistent with prevOutScript')
 
-    expanded = expandOutput(witnessScript, undefined, kpPubKey)
-    if (!expanded.pubKeys) throw new Error('WitnessScript not supported "' + bscript.toASM(redeemScript) + '"')
-
-    prevOutType = btemplates.types.P2SH
-    prevOutScript = btemplates.scriptHash.output.encode(redeemScriptHash)
-    p2sh = witness = p2wsh = true
-    p2shType = btemplates.types.P2WSH
-    signType = witnessType = expanded.scriptType
-    signScript = witnessScript
-  } else if (redeemScript) {
-    redeemScriptHash = bcrypto.hash160(redeemScript)
-    checkP2SHInput(input, redeemScriptHash)
-
-    expanded = expandOutput(redeemScript, undefined, kpPubKey)
-    if (!expanded.pubKeys) throw new Error('RedeemScript not supported "' + bscript.toASM(redeemScript) + '"')
-
-    prevOutType = btemplates.types.P2SH
-    prevOutScript = btemplates.scriptHash.output.encode(redeemScriptHash)
-    p2sh = true
-    signType = p2shType = expanded.scriptType
-    signScript = redeemScript
-    witness = signType === btemplates.types.P2WPKH
-  } else if (witnessScript) {
-    witnessScriptHash = bcrypto.sha256(witnessScript)
-    checkP2WSHInput(input, witnessScriptHash)
-
-    expanded = expandOutput(witnessScript, undefined, kpPubKey)
-    if (!expanded.pubKeys) throw new Error('WitnessScript not supported "' + bscript.toASM(redeemScript) + '"')
-
-    prevOutType = btemplates.types.P2WSH
-    prevOutScript = btemplates.witnessScriptHash.output.encode(witnessScriptHash)
-    witness = p2wsh = true
-    signType = witnessType = expanded.scriptType
-    signScript = witnessScript
-  } else if (input.prevOutType) {
-    // embedded scripts are not possible without a redeemScript
-    if (input.prevOutType === scriptTypes.P2SH ||
-      input.prevOutType === scriptTypes.P2WSH) {
-      throw new Error('PrevOutScript is ' + input.prevOutType + ', requires redeemScript')
+    const expanded = expandOutput(p2wsh.redeem.output, ourPubKey)
+    if (!expanded.pubkeys) throw new Error(expanded.type + ' not supported as witnessScript (' + bscript.toASM(witnessScript) + ')')
+    if (input.signatures && input.signatures.some(x => x)) {
+      expanded.signatures = input.signatures
     }
 
-    prevOutType = input.prevOutType
-    prevOutScript = input.prevOutScript
-    expanded = expandOutput(input.prevOutScript, input.prevOutType, kpPubKey)
-    if (!expanded.pubKeys) return
+    let signScript = witnessScript
+    if (expanded.type === SCRIPT_TYPES.P2WPKH) throw new Error('P2SH(P2WSH(P2WPKH)) is a consensus failure')
 
-    witness = (input.prevOutType === scriptTypes.P2WPKH)
-    signType = prevOutType
-    signScript = prevOutScript
-  } else {
-    prevOutScript = btemplates.pubKeyHash.output.encode(bcrypto.hash160(kpPubKey))
-    expanded = expandOutput(prevOutScript, scriptTypes.P2PKH, kpPubKey)
+    return {
+      redeemScript,
+      redeemScriptType: SCRIPT_TYPES.P2WSH,
 
-    prevOutType = scriptTypes.P2PKH
-    witness = false
-    signType = prevOutType
-    signScript = prevOutScript
-  }
+      witnessScript,
+      witnessScriptType: expanded.type,
 
-  if (signType === scriptTypes.P2WPKH) {
-    signScript = btemplates.pubKeyHash.output.encode(btemplates.witnessPubKeyHash.output.decode(signScript))
-  }
+      prevOutType: SCRIPT_TYPES.P2SH,
+      prevOutScript: p2sh.output,
 
-  if (p2sh) {
-    input.redeemScript = redeemScript
-    input.redeemScriptType = p2shType
-  }
+      hasWitness: true,
+      signScript,
+      signType: expanded.type,
 
-  if (p2wsh) {
-    input.witnessScript = witnessScript
-    input.witnessScriptType = witnessType
-  }
-
-  input.pubKeys = expanded.pubKeys
-  input.signatures = expanded.signatures
-  input.signScript = signScript
-  input.signType = signType
-  input.prevOutScript = prevOutScript
-  input.prevOutType = prevOutType
-  input.witness = witness
-}
-
-function buildStack (type, signatures, pubKeys, allowIncomplete) {
-  if (type === scriptTypes.P2PKH) {
-    if (signatures.length === 1 && Buffer.isBuffer(signatures[0]) && pubKeys.length === 1) return btemplates.pubKeyHash.input.encodeStack(signatures[0], pubKeys[0])
-  } else if (type === scriptTypes.P2PK) {
-    if (signatures.length === 1 && Buffer.isBuffer(signatures[0])) return btemplates.pubKey.input.encodeStack(signatures[0])
-  } else if (type === scriptTypes.MULTISIG) {
-    if (signatures.length > 0) {
-      signatures = signatures.map(function (signature) {
-        return signature || ops.OP_0
-      })
-      if (!allowIncomplete) {
-        // remove blank signatures
-        signatures = signatures.filter(function (x) { return x !== ops.OP_0 })
-      }
-
-      return btemplates.multisig.input.encodeStack(signatures)
-    }
-  } else {
-    throw new Error('Not yet supported')
-  }
-
-  if (!allowIncomplete) throw new Error('Not enough signatures provided')
-  return []
-}
-
-function buildInput (input, allowIncomplete) {
-  var scriptType = input.prevOutType
-  var sig = []
-  var witness = []
-
-  if (supportedType(scriptType)) {
-    sig = buildStack(scriptType, input.signatures, input.pubKeys, allowIncomplete)
-  }
-
-  var p2sh = false
-  if (scriptType === btemplates.types.P2SH) {
-    // We can remove this error later when we have a guarantee prepareInput
-    // rejects unsignable scripts - it MUST be signable at this point.
-    if (!allowIncomplete && !supportedP2SHType(input.redeemScriptType)) {
-      throw new Error('Impossible to sign this type')
-    }
-
-    if (supportedType(input.redeemScriptType)) {
-      sig = buildStack(input.redeemScriptType, input.signatures, input.pubKeys, allowIncomplete)
-    }
-
-    // If it wasn't SIGNABLE, it's witness, defer to that
-    if (input.redeemScriptType) {
-      p2sh = true
-      scriptType = input.redeemScriptType
+      pubkeys: expanded.pubkeys,
+      signatures: expanded.signatures
     }
   }
 
-  switch (scriptType) {
-    // P2WPKH is a special case of P2PKH
-    case btemplates.types.P2WPKH:
-      witness = buildStack(btemplates.types.P2PKH, input.signatures, input.pubKeys, allowIncomplete)
-      break
+  if (redeemScript) {
+    const p2sh = payments.p2sh({ redeem: { output: redeemScript } })
 
-    case btemplates.types.P2WSH:
-      // We can remove this check later
-      if (!allowIncomplete && !supportedType(input.witnessScriptType)) {
-        throw new Error('Impossible to sign this type')
-      }
+    if (input.prevOutScript) {
+      let p2shAlt
+      try {
+        p2shAlt = payments.p2sh({ output: input.prevOutScript })
+      } catch (e) { throw new Error('PrevOutScript must be P2SH') }
+      if (!p2sh.hash.equals(p2shAlt.hash)) throw new Error('Redeem script inconsistent with prevOutScript')
+    }
 
-      if (supportedType(input.witnessScriptType)) {
-        witness = buildStack(input.witnessScriptType, input.signatures, input.pubKeys, allowIncomplete)
-        witness.push(input.witnessScript)
-        scriptType = input.witnessScriptType
-      }
+    const expanded = expandOutput(p2sh.redeem.output, ourPubKey)
+    if (!expanded.pubkeys) throw new Error(expanded.type + ' not supported as redeemScript (' + bscript.toASM(redeemScript) + ')')
+    if (input.signatures && input.signatures.some(x => x)) {
+      expanded.signatures = input.signatures
+    }
 
-      break
+    let signScript = redeemScript
+    if (expanded.type === SCRIPT_TYPES.P2WPKH) {
+      signScript = payments.p2pkh({ pubkey: expanded.pubkeys[0] }).output
+    }
+
+    return {
+      redeemScript,
+      redeemScriptType: expanded.type,
+
+      prevOutType: SCRIPT_TYPES.P2SH,
+      prevOutScript: p2sh.output,
+
+      hasWitness: expanded.type === SCRIPT_TYPES.P2WPKH,
+      signScript,
+      signType: expanded.type,
+
+      pubkeys: expanded.pubkeys,
+      signatures: expanded.signatures
+    }
   }
 
-  // append redeemScript if necessary
-  if (p2sh) {
-    sig.push(input.redeemScript)
+  if (witnessScript) {
+    const p2wsh = payments.p2wsh({ redeem: { output: witnessScript } })
+
+    if (input.prevOutScript) {
+      const p2wshAlt = payments.p2wsh({ output: input.prevOutScript })
+      if (!p2wsh.hash.equals(p2wshAlt.hash)) throw new Error('Witness script inconsistent with prevOutScript')
+    }
+
+    const expanded = expandOutput(p2wsh.redeem.output, ourPubKey)
+    if (!expanded.pubkeys) throw new Error(expanded.type + ' not supported as witnessScript (' + bscript.toASM(witnessScript) + ')')
+    if (input.signatures && input.signatures.some(x => x)) {
+      expanded.signatures = input.signatures
+    }
+
+    let signScript = witnessScript
+    if (expanded.type === SCRIPT_TYPES.P2WPKH) throw new Error('P2WSH(P2WPKH) is a consensus failure')
+
+    return {
+      witnessScript,
+      witnessScriptType: expanded.type,
+
+      prevOutType: SCRIPT_TYPES.P2WSH,
+      prevOutScript: p2wsh.output,
+
+      hasWitness: true,
+      signScript,
+      signType: expanded.type,
+
+      pubkeys: expanded.pubkeys,
+      signatures: expanded.signatures
+    }
   }
 
+  if (input.prevOutType && input.prevOutScript) {
+    // embedded scripts are not possible without extra information
+    if (input.prevOutType === SCRIPT_TYPES.P2SH) throw new Error('PrevOutScript is ' + input.prevOutType + ', requires redeemScript')
+    if (input.prevOutType === SCRIPT_TYPES.P2WSH) throw new Error('PrevOutScript is ' + input.prevOutType + ', requires witnessScript')
+    if (!input.prevOutScript) throw new Error('PrevOutScript is missing')
+
+    const expanded = expandOutput(input.prevOutScript, ourPubKey)
+    if (!expanded.pubkeys) throw new Error(expanded.type + ' not supported (' + bscript.toASM(input.prevOutScript) + ')')
+    if (input.signatures && input.signatures.some(x => x)) {
+      expanded.signatures = input.signatures
+    }
+
+    let signScript = input.prevOutScript
+    if (expanded.type === SCRIPT_TYPES.P2WPKH) {
+      signScript = payments.p2pkh({ pubkey: expanded.pubkeys[0] }).output
+    }
+
+    return {
+      prevOutType: expanded.type,
+      prevOutScript: input.prevOutScript,
+
+      hasWitness: expanded.type === SCRIPT_TYPES.P2WPKH,
+      signScript,
+      signType: expanded.type,
+
+      pubkeys: expanded.pubkeys,
+      signatures: expanded.signatures
+    }
+  }
+
+  const prevOutScript = payments.p2pkh({ pubkey: ourPubKey }).output
   return {
-    type: scriptType,
-    script: bscript.compile(sig),
-    witness: witness
+    prevOutType: SCRIPT_TYPES.P2PKH,
+    prevOutScript: prevOutScript,
+
+    hasWitness: false,
+    signScript: prevOutScript,
+    signType: SCRIPT_TYPES.P2PKH,
+
+    pubkeys: [ourPubKey],
+    signatures: [undefined]
+  }
+}
+
+function build (type, input, allowIncomplete) {
+  const pubkeys = input.pubkeys || []
+  let signatures = input.signatures || []
+
+  switch (type) {
+    case SCRIPT_TYPES.P2PKH: {
+      if (pubkeys.length === 0) break
+      if (signatures.length === 0) break
+
+      return payments.p2pkh({ pubkey: pubkeys[0], signature: signatures[0] })
+    }
+    case SCRIPT_TYPES.P2WPKH: {
+      if (pubkeys.length === 0) break
+      if (signatures.length === 0) break
+
+      return payments.p2wpkh({ pubkey: pubkeys[0], signature: signatures[0] })
+    }
+    case SCRIPT_TYPES.P2PK: {
+      if (pubkeys.length === 0) break
+      if (signatures.length === 0) break
+
+      return payments.p2pk({ signature: signatures[0] })
+    }
+    case SCRIPT_TYPES.MULTISIG: {
+      if (allowIncomplete) {
+        signatures = signatures.map(x => x || ops.OP_0)
+      } else {
+        signatures = signatures.filter(x => x)
+      }
+
+      return payments.p2ms({ signatures }, { allowIncomplete })
+    }
+    case SCRIPT_TYPES.P2SH: {
+      const redeem = build(input.redeemScriptType, input, allowIncomplete)
+      if (!redeem) return
+
+      return payments.p2sh({
+        redeem: {
+          output: redeem.output || input.redeemScript,
+          input: redeem.input,
+          witness: redeem.witness
+        }
+      })
+    }
+    case SCRIPT_TYPES.P2WSH: {
+      const redeem = build(input.witnessScriptType, input, allowIncomplete)
+      if (!redeem) return
+
+      return payments.p2wsh({
+        redeem: {
+          output: input.witnessScript,
+          input: redeem.input,
+          witness: redeem.witness
+        }
+      })
+    }
   }
 }
 
 function TransactionBuilder (network, maximumFeeRate) {
-  this.prevTxMap = {}
+  this.__prevTxSet = {}
   this.network = network || networks.bitcoin
 
   // WARNING: This is __NOT__ to be relied on, its just another potential safety mechanism (safety in-depth)
   this.maximumFeeRate = maximumFeeRate || 2500
 
-  this.inputs = []
-  this.tx = new Transaction()
+  this.__inputs = []
+  this.__tx = new Transaction()
+  this.__tx.version = 2
 }
 
 TransactionBuilder.prototype.setLockTime = function (locktime) {
   typeforce(types.UInt32, locktime)
 
   // if any signatures exist, throw
-  if (this.inputs.some(function (input) {
+  if (this.__inputs.some(function (input) {
     if (!input.signatures) return false
 
     return input.signatures.some(function (s) { return s })
@@ -488,18 +453,18 @@ TransactionBuilder.prototype.setLockTime = function (locktime) {
     throw new Error('No, this would invalidate signatures')
   }
 
-  this.tx.locktime = locktime
+  this.__tx.locktime = locktime
 }
 
 TransactionBuilder.prototype.setVersion = function (version) {
   typeforce(types.UInt32, version)
 
   // XXX: this might eventually become more complex depending on what the versions represent
-  this.tx.version = version
+  this.__tx.version = version
 }
 
 TransactionBuilder.fromTransaction = function (transaction, network) {
-  var txb = new TransactionBuilder(network)
+  const txb = new TransactionBuilder(network)
 
   // Copy transaction fields
   txb.setVersion(transaction.version)
@@ -520,7 +485,7 @@ TransactionBuilder.fromTransaction = function (transaction, network) {
   })
 
   // fix some things not possible through the public API
-  txb.inputs.forEach(function (input, i) {
+  txb.__inputs.forEach(function (input, i) {
     fixMultisigOrder(input, transaction, i)
   })
 
@@ -532,16 +497,16 @@ TransactionBuilder.prototype.addInput = function (txHash, vout, sequence, prevOu
     throw new Error('No, this would invalidate signatures')
   }
 
-  var value
+  let value
 
   // is it a hex string?
   if (typeof txHash === 'string') {
     // transaction hashs's are displayed in reverse order, un-reverse it
-    txHash = bufferReverse(new Buffer(txHash, 'hex'))
+	txHash = bufferReverse(new Buffer(txHash, 'hex'))
 
   // is it a Transaction object?
   } else if (txHash instanceof Transaction) {
-    var txOut = txHash.outs[vout]
+    const txOut = txHash.outs[vout]
     prevOutScript = txOut.script
     value = txOut.value
 
@@ -560,10 +525,10 @@ TransactionBuilder.prototype.__addInputUnsafe = function (txHash, vout, options)
     throw new Error('coinbase inputs not supported')
   }
 
-  var prevTxOut = txHash.toString('hex') + ':' + vout
-  if (this.prevTxMap[prevTxOut] !== undefined) throw new Error('Duplicate TxOut: ' + prevTxOut)
+  const prevTxOut = txHash.toString('hex') + ':' + vout
+  if (this.__prevTxSet[prevTxOut] !== undefined) throw new Error('Duplicate TxOut: ' + prevTxOut)
 
-  var input = {}
+  let input = {}
 
   // derive what we can from the scriptSig
   if (options.script !== undefined) {
@@ -577,26 +542,25 @@ TransactionBuilder.prototype.__addInputUnsafe = function (txHash, vout, options)
 
   // derive what we can from the previous transactions output script
   if (!input.prevOutScript && options.prevOutScript) {
-    var prevOutType
+    let prevOutType
 
-    if (!input.pubKeys && !input.signatures) {
-      var expanded = expandOutput(options.prevOutScript)
-
-      if (expanded.pubKeys) {
-        input.pubKeys = expanded.pubKeys
+    if (!input.pubkeys && !input.signatures) {
+      const expanded = expandOutput(options.prevOutScript)
+      if (expanded.pubkeys) {
+        input.pubkeys = expanded.pubkeys
         input.signatures = expanded.signatures
       }
 
-      prevOutType = expanded.scriptType
+      prevOutType = expanded.type
     }
 
     input.prevOutScript = options.prevOutScript
-    input.prevOutType = prevOutType || btemplates.classifyOutput(options.prevOutScript)
+    input.prevOutType = prevOutType || classify.output(options.prevOutScript)
   }
 
-  var vin = this.tx.addInput(txHash, vout, options.sequence, options.scriptSig)
-  this.inputs[vin] = input
-  this.prevTxMap[prevTxOut] = vin
+  const vin = this.__tx.addInput(txHash, vout, options.sequence, options.scriptSig)
+  this.__inputs[vin] = input
+  this.__prevTxSet[prevTxOut] = true
   return vin
 }
 
@@ -605,12 +569,12 @@ TransactionBuilder.prototype.addOutput = function (scriptPubKey, value) {
     throw new Error('No, this would invalidate signatures')
   }
 
-  // Attempt to get a script if it's a base58 address string
+  // Attempt to get a script if it's a base58 or bech32 address string
   if (typeof scriptPubKey === 'string') {
     scriptPubKey = baddress.toOutputScript(scriptPubKey, this.network)
   }
 
-  return this.tx.addOutput(scriptPubKey, value)
+  return this.__tx.addOutput(scriptPubKey, value)
 }
 
 TransactionBuilder.prototype.build = function () {
@@ -622,25 +586,24 @@ TransactionBuilder.prototype.buildIncomplete = function () {
 
 TransactionBuilder.prototype.__build = function (allowIncomplete) {
   if (!allowIncomplete) {
-    if (!this.tx.ins.length) throw new Error('Transaction has no inputs')
-    if (!this.tx.outs.length) throw new Error('Transaction has no outputs')
+    if (!this.__tx.ins.length) throw new Error('Transaction has no inputs')
+    if (!this.__tx.outs.length) throw new Error('Transaction has no outputs')
   }
 
-  var tx = this.tx.clone()
-  // Create script signatures from inputs
-  this.inputs.forEach(function (input, i) {
-    var scriptType = input.witnessScriptType || input.redeemScriptType || input.prevOutType
-    if (!scriptType && !allowIncomplete) throw new Error('Transaction is not complete')
-    var result = buildInput(input, allowIncomplete)
+  const tx = this.__tx.clone()
 
-    // skip if no result
-    if (!allowIncomplete) {
-      if (!supportedType(result.type) && result.type !== btemplates.types.P2WPKH) {
-        throw new Error(result.type + ' not supported')
-      }
+  // create script signatures from inputs
+  this.__inputs.forEach(function (input, i) {
+    if (!input.prevOutType && !allowIncomplete) throw new Error('Transaction is not complete')
+
+    const result = build(input.prevOutType, input, allowIncomplete)
+    if (!result) {
+      if (!allowIncomplete && input.prevOutType === SCRIPT_TYPES.NONSTANDARD) throw new Error('Unknown input type')
+      if (!allowIncomplete) throw new Error('Not enough information')
+      return
     }
 
-    tx.setInputScript(i, result.script)
+    tx.setInputScript(i, result.input)
     tx.setWitness(i, result.witness)
   })
 
@@ -655,25 +618,25 @@ TransactionBuilder.prototype.__build = function (allowIncomplete) {
 }
 
 function canSign (input) {
-  return input.prevOutScript !== undefined &&
-    input.signScript !== undefined &&
-    input.pubKeys !== undefined &&
+  return input.signScript !== undefined &&
+    input.signType !== undefined &&
+    input.pubkeys !== undefined &&
     input.signatures !== undefined &&
-    input.signatures.length === input.pubKeys.length &&
-    input.pubKeys.length > 0 &&
+    input.signatures.length === input.pubkeys.length &&
+    input.pubkeys.length > 0 &&
     (
-      input.witness === false ||
-      (input.witness === true && input.value !== undefined)
+      input.hasWitness === false ||
+      input.value !== undefined
     )
 }
 
 TransactionBuilder.prototype.sign = function (vin, keyPair, redeemScript, hashType, witnessValue, witnessScript) {
   // TODO: remove keyPair.network matching in 4.0.0
   if (keyPair.network && keyPair.network !== this.network) throw new TypeError('Inconsistent network')
-  if (!this.inputs[vin]) throw new Error('No input at index: ' + vin)
+  if (!this.__inputs[vin]) throw new Error('No input at index: ' + vin)
   hashType = hashType || Transaction.SIGHASH_ALL
 
-  var input = this.inputs[vin]
+  const input = this.__inputs[vin]
 
   // if redeemScript was previously provided, enforce consistency
   if (input.redeemScript !== undefined &&
@@ -682,7 +645,7 @@ TransactionBuilder.prototype.sign = function (vin, keyPair, redeemScript, hashTy
     throw new Error('Inconsistent redeemScript')
   }
 
-  var kpPubKey = keyPair.publicKey || keyPair.getPublicKeyBuffer()
+  const ourPubKey = keyPair.publicKey || keyPair.getPublicKey()
   if (!canSign(input)) {
     if (witnessValue !== undefined) {
       if (input.value !== undefined && input.value !== witnessValue) throw new Error('Input didn\'t match witnessValue')
@@ -690,29 +653,36 @@ TransactionBuilder.prototype.sign = function (vin, keyPair, redeemScript, hashTy
       input.value = witnessValue
     }
 
-    if (!canSign(input)) prepareInput(input, kpPubKey, redeemScript, witnessValue, witnessScript)
+    if (!canSign(input)) {
+      const prepared = prepareInput(input, ourPubKey, redeemScript, witnessValue, witnessScript)
+
+      // updates inline
+      Object.assign(input, prepared)
+    }
+
     if (!canSign(input)) throw Error(input.prevOutType + ' not supported')
   }
 
   // ready to sign
-  var signatureHash
-  if (input.witness) {
-    signatureHash = this.tx.hashForWitnessV0(vin, input.signScript, input.value, hashType)
+  let signatureHash
+  if (input.hasWitness) {
+    signatureHash = this.__tx.hashForWitnessV0(vin, input.signScript, input.value, hashType)
   } else {
-    signatureHash = this.tx.hashForSignature(vin, input.signScript, hashType)
+    signatureHash = this.__tx.hashForSignature(vin, input.signScript, hashType)
   }
 
   // enforce in order signing of public keys
-  var signed = input.pubKeys.some(function (pubKey, i) {
-    if (!kpPubKey.equals(pubKey)) return false
+  const signed = input.pubkeys.some(function (pubKey, i) {
+    if (!ourPubKey.equals(pubKey)) return false
     if (input.signatures[i]) throw new Error('Signature already exists')
-    if (kpPubKey.length !== 33 &&
-      input.signType === scriptTypes.P2WPKH) throw new Error('BIP143 rejects uncompressed public keys in P2WPKH or P2WSH')
 
-    var signature = keyPair.sign(signatureHash)
-    if (Buffer.isBuffer(signature)) signature = ECSignature.fromRSBuffer(signature)
+    // TODO: add tests
+    if (ourPubKey.length !== 33 && input.hasWitness) {
+      throw new Error('BIP143 rejects uncompressed public keys in P2WPKH or P2WSH')
+    }
 
-    input.signatures[i] = signature.toScriptSignature(hashType)
+    const signature = keyPair.sign(signatureHash)
+    input.signatures[i] = bscript.signature.encode(signature, hashType)
     return true
   })
 
@@ -724,13 +694,13 @@ function signatureHashType (buffer) {
 }
 
 TransactionBuilder.prototype.__canModifyInputs = function () {
-  return this.inputs.every(function (input) {
+  return this.__inputs.every(function (input) {
     // any signatures?
     if (input.signatures === undefined) return true
 
     return input.signatures.every(function (signature) {
       if (!signature) return true
-      var hashType = signatureHashType(signature)
+      const hashType = signatureHashType(signature)
 
       // if SIGHASH_ANYONECANPAY is set, signatures would not
       // be invalidated by more inputs
@@ -740,17 +710,17 @@ TransactionBuilder.prototype.__canModifyInputs = function () {
 }
 
 TransactionBuilder.prototype.__canModifyOutputs = function () {
-  var nInputs = this.tx.ins.length
-  var nOutputs = this.tx.outs.length
+  const nInputs = this.__tx.ins.length
+  const nOutputs = this.__tx.outs.length
 
-  return this.inputs.every(function (input) {
+  return this.__inputs.every(function (input) {
     if (input.signatures === undefined) return true
 
     return input.signatures.every(function (signature) {
       if (!signature) return true
-      var hashType = signatureHashType(signature)
+      const hashType = signatureHashType(signature)
 
-      var hashTypeMod = hashType & 0x1f
+      const hashTypeMod = hashType & 0x1f
       if (hashTypeMod === Transaction.SIGHASH_NONE) return true
       if (hashTypeMod === Transaction.SIGHASH_SINGLE) {
         // if SIGHASH_SINGLE is set, and nInputs > nOutputs
@@ -764,13 +734,13 @@ TransactionBuilder.prototype.__canModifyOutputs = function () {
 
 TransactionBuilder.prototype.__overMaximumFees = function (bytes) {
   // not all inputs will have .value defined
-  var incoming = this.inputs.reduce(function (a, x) { return a + (x.value >>> 0) }, 0)
+  const incoming = this.__inputs.reduce(function (a, x) { return a + (x.value >>> 0) }, 0)
 
   // but all outputs do, and if we have any input value
   // we can immediately determine if the outputs are too small
-  var outgoing = this.tx.outs.reduce(function (a, x) { return a + x.value }, 0)
-  var fee = incoming - outgoing
-  var feeRate = fee / bytes
+  const outgoing = this.__tx.outs.reduce(function (a, x) { return a + x.value }, 0)
+  const fee = incoming - outgoing
+  const feeRate = fee / bytes
 
   return feeRate > this.maximumFeeRate
 }
